Working on the http layer, adding 'create' route
Many things here are not ideal. There is very little checking that's
being done, for this we need to tweak the core API a bit further. There
are some corner cases that aren't caught. Authentication isn't being
done yet. And it's not 100% clear yet where those scopes should live.

But...there is progress being made. The API will develop and hopefully
soon-ish we'll also be able to add versioning via a new serde-versioning
crate being written (hint hint).

The basic problem is that we need to build a system that works for two kinds
of settings: remotely, administered via a few users, on a different computer
with limited access and local, used by a single user and application stack
(http interface for browser plugins, file API for everything else).

The latter is the focus of development for now, but before we hit 1.0 (aka
stable for production) there are more issues to be resolved.

Also...if lockchain is to be used as a keystore in poke, we need to figure
out how to allow easier integration of the core storage components into other
applications (maybe some guides).

Also...all of this should be documented in other places than just my ramblings
in commit messages 😅