Starting work on Keyfold
The idea here is the following:

You don't want to have to write encryption code for keys all the time
especially if there are many and maybe lots of interactions.

Also...maybe you doN't have the raw key anymore. It's kept in the vault
in encrypted form and you just cache it from time to time. So you get it,
decrypt it, store it in the Engine again, do some crypto. The user forgets
about you, the key is cleared..etc.

The keyfold comes in right there! It takes care of decrypting and
re-encrypting any keychanges made by the user, so that the Vault
backend only ever gets encrypted keys to work with.