Changing license
Initially I decided lockchain should be an MIT licensed piece of
software to allow a wide range of people and projects to use
it. Thinking about it now, I feel that there are enough proprietary
solutions for secret storage in the world. Whether that be pure
password managers or enterprise-focused distributed secret stores.

I don't know what lockchain will be one day or even what it is today,
but I do know that I don't want it to become the backend to a
proprietary piece of software. Therefore I have decided to relicense
it under the GNU General Public License 3.0 or later.